move SummaryCard to components dir
